breast_cancer_system/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy
pymysql
scikit-learn
pandas
python-jose[cryptography]
python-multipart
pytest
